Final Rev 1 version, tested and working
Has PID values preset and setpoint at 100

+++ src/pidtask.h
@@ -15,12 +15,12 @@
 
 /* PID structure definition */
 typedef struct {
-    uint32_t Kp;           // proportional gain
+    float Kp;           // proportional gain
     float Ki;           // integral gain
     float Kd;           // derivative gain
     uint16_t setpoint;  // intensity setpoint
     float integral;     // error acccumulator, sum of errors over time
-    uint16_t prev_error;   // previous error, used for (d_error/dt)
+    float prev_error;   // previous error, used for (d_error/dt)
     float delta_t;      // change in time for derivative
     uint16_t max_lim;      // value to set upper rail for output
     uint8_t min_lim;      // value to set lower rail for output
